Tests Added for MoreQuestionnaireItemComponent Class
